Set row heights based on pixels rather than points
Previously, on a high-DPI display, rows in the ctlSQLGrid
would be too small by default.  Although the user can
manually adjust row heights, this is tedious and makes it
almost impossible to skim through data.  By calculating
the default row height based on pixels rather than points,
this is taken into account.

There are still some magic numbers here.  Using
GetPixelSize() alone seems to place the baseline of the
text at the bottom of the cell, cutting off any descenders
in the text.  To adjust for this, we add 20% of one line
height (experimentally determined).  Unfortunately, we
can't use wxDC::GetTextExtent, which would give us precise
results, because we don't have a wxDC at the time we're
setting these default heights.